scripts/requirements.txt
boto3==1.34.34
botocore==1.34.34
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
